mixins now do not have priority over inheritance
Before this change the mixins were only created at instantiation time:
all the mixins were first collected, then a new temporary class with
the mixins and the actual class was created and instantiated.

Now the mixins alter the class which they are added to. Subclassing
works as if the base class with mixins was subclassed.

This is what developers intuitively expect. Mixins modify classes they
are attached to.