(no ticket) Specified versions of beta dependencies in setup.py
The lastest version of pip refuses to install pre-release versions
of dependencies (if not run with --pre), even if stable are not
available. Unfortunately, we need some packages which pip recognizes
as pre- even if they are not (because they use non-standard versioning
conventions). Fortunately, we can fix it by explicitly specifying
pre- versions in setup.py.

Change-Id: Ia0498fc844eb699c851813e92de0da4dbbb36db4