(no ticket) Bumped celery to >=3.1.15
This fixes unsolvable package version requirements, namely
librabbitmq requiring amqp>=1.2.1 and celery<3.1 forcing amqp<1.2.

I did a simple test with a local Celery instance without RabbitMQ.
No problems seen.

There is no reason stated in the change introducing this celery<3.1
dependency, but maybe it was related to the fact that django-celery
may had not been ready when Celery 3.1 was released.

Change-Id: I16ecd8f8926b452ef1eb184cbbd00e36f9b600e1