(no-ticket) Explicitly silence some NoReverseMatch exceptions
In Django 1.5, when a NoReverseMatch exception was raised from a method
when rendering a template, it was silenced and the method returned
"(None)". It is no longer being silenced in Django 1.6, so we must
do it explicitly.
https://docs.djangoproject.com/en/1.6/releases/1.6/#miscellaneous
The "Upgrade to Django 1.6" commit fixed this partially, but some places
where NoReverseMatch was thrown were missed.

Change-Id: I230a9d809c1133c01716ee621348d4d2a5ce6a1c