(no-ticket) Fix a race condition in the add_or_update_problem view
In the add_or_update_problem view, before unpackmgr is fired to unpack
the problem package, changes to the database are made. Previously
the whole view was wrapped inside a transaction, so if unpackmgr ran
before the view finished and the changes were commited, it didn't
see the changes and bad things happened. This change makes the whole
view non-atomic but takes care to wrap some necessary parts in
a transaction.

Change-Id: I0dfd511854b67f0b0b2bb346d354309e53d9a2f0