SIO-1641 Contest-free side of SIO2.
The current contest mechanism has been overhauled completely.
The active contest isn't being held in the session anymore. A contest
is active if and only if the visited URL is prefixed with the contest's
ID.
Administrators can use a setting to decide if users should be put
inside a contest forcibly (which happens by redirections) or if they can
visit pages without participating in any contest, and in particular, if
they can visit pages that require no contest to be active.
Developers can write views that require a contest, that are neutral, or
that require no contest to be active. The decision is made in the
urlconf by placing patterns pointing to the views in one of three
variables. Also ModelAdmins can be registered in two different sites,
where one of them contains ModelAdmins that require a contest.
Read the docs for more information.

Change-Id: Iea9640060b9b580ae0bd76aafd987522e9fcddb6