(no-ticket) Fix non contest submission views.
+ extend the modified reverse API so it allows asking for non-contest URL
explicitly (it was possible before, but not documented and hacky). Also,
when explicitly asking for a contest URL, if one doesn't exist then
reversing will always fail (instead of returning a non-contest version
or the global version if one exists).
+ extend the tabbed_view function to allow request-based conditions
for accessing the tabs
+ stop showing the "Submissions" and "Submit" tabs in the problem site
if there is a currently active contest. Submitting to contest should be
done from the contest submission page. Submitting outside any contest
should be done when no contest is active.
+ little fix in the settings template

Change-Id: Ibe9729c440ffb8312c1069fe97eb8194de42f73e