(no-ticket) Prettify portal problemtable. Show empty tables.
Column widths in portal ProblemTable widget are now fixed so that consecutive
tables are aligned and generally prettier.

ProblemTable markup tags are now properly parsed and hidden instead of
appearing in plaintext if they contain no tasks.

Layout is prettified, also in addition to score link to submissin is displayed.

Change-Id: Iafd96d2a1b8175b06061ab04ca1219da787fed6e