SIO-1912 Return the real message ID if the user is not allowed to see the top reference
Before this change we always returned the top reference. It was a problem
when a public reply was created and most users did not have the
permissions to see the original message. (only the reply is public)

It would be intuitive to return the real message ID always, but
it's false intution. Then only the message would be displayed and
the parent message - never, even if the user has rights to see it.

Thus, when sending e-mails we check if the user has permissions to
see the top reference and, if so, send it. Otherwise we return the
real message ID, to be on the safe side.

Change-Id: Id4b0777fb8aaa6f9d8e675a8b9d36ce887a50ea0