(no-ticket) Fix queuemanager to proper receive notifications.
I removed this line by mistake when I was fixing notifications module.
It is crucial to call close on the socket and then manually delete its entry
from dictionary. If we delete entry by subscribing on socket to close call
we could try to close it twice before entry would be deleted.

Change-Id: Ic43c49a724dec7a482b293f425f22595e09b870d