(SIO-1823) Backend for enabling multilingual problem statements
This change allows to upload multiple problem statements with one package.
By default allowed addition languages are endings -pl and -en, but they can be
adjusted in settings.py.
Note that this change does not allow to browse these files at will just yet.
However, if we switch to english version of the site, it will open english
version of problem statement.

Change-Id: I85f642ff5382f7da449621877809ce103fb98cff