(no-ticket) Fix submission spam when creating model solutions
When uploading a package, model solutions are created one by one.
However, with each model solution created, all previously existing
model solutions are rejudged. That means that for n modelsolutions,
there's O(n^2) rejudges.

What's worse, when rejudging a model solution, the previous submission
for it is immediately deleted. If this happens before the sumbission
reaches evalmgr, evalmgr will think it's not in the database _yet_
(as opposed to not being in database _anymore_) and will keep sleeping
1 second and re-checking if it maybe appeared in the database.
It will give up after a few checks, but this happens for all but the last
submission for each model solution, which means it happens O(n^2) times.
This blocks evalmgr for an unacceptable amount of time.

Fix this by only judging the modelsolution we're currently adding,
if we're adding it as a result of package upload.

Change-Id: I937366c827cd4395f9b60625b1e22d3f251820a5