SIO-1926 Fix requirements.txt version syntax
For some reason, the #egg=name-version syntax does not work as stated
in pip documentation, producing egg_info warnings and causing the
Dockerfile build to fail.
What seems to work as intended is #egg=name==version - running pip
install with verbose option confirms that it actually checks if the
downloaded package version is correct.

Change-Id: I735aec1f7cdddbce0ed676f62fe8d3eeb8450c06