(no-ticket) Don't chown if owner is already correct
Sometimes running recursive chown in a docker container is very slow.
(see https://github.com/docker/for-linux/issues/388)
The files should have been added to the container with the correct
owner, but the mounted volume's contents may have been modified after
building the docker image. It is possible, and much faster, to use find
to only chown the files, whose owner is not already correct.

Change-Id: Ic0313cd02c70f77683204e5c5973c634ebc736fd