(no-ticket) Fix check if user already has a team
Previous check if a TeamMembership is valid queries database for
TeamMemberships with same user and same contest excluding same team.
This checks fails to work as expected when team was not commited to
database due to https://code.djangoproject.com/ticket/25467. After this
change we only exclude the team if it was commited to database. Also
added a test for found problem.

Change-Id: I06b04c997c9881db7d02bd17cc44f0a420118ffe