SIO-2134 Review options visible in contest admins
There are functions has_add(/change/delete)_premmission which
should be added to class representing item in admin.py wherever
to affect it visibility. They were missing in several places and
in others just returned True or checked premissions guaranted by
adding module to proper admin object.

Now the functions are added to super class and by default return
True as it is most common intention when adding an item to menu.

Change-Id: Iab9a8d4eedb242f618c2bbebcfda6af5d163a3d3