(SIO-2168) Set on_delete behaviour
This solved following warning

RemovedInDjango20Warning: on_delete will be a required arg for *** in Django 2.0. Set it to models.CASCADE on models and in existing migrations if you want to maintain the current default behavior.

Behaviour is set to the current default of CASCADE where undefined.

This commit doesn't introduce any difference in code logic.

Change-Id: I8439d40e0d5a6b16336a8f4ab4983f9e7fc52e08