(no-ticket) Fix bugs not caught due to never executed tests
Previously, @needs_linux decorator was introduced to limit execution of
sinol makefile related tests to Linux only. It had a side effect of
skipping execution of those tests entirely, even on Linux.

Tests in question had to be fixed because no one noticed them for
so long. A small number of bugs was discovered throught this effort:

 - addproblem management command not starting since Django 1.10 upgrade
   due to argument parsing failure,
 - assertStreamingEqual sometimes has to let PDFs pass, therefore
   decoding what passes through it is unwise - this behaviour was
   introduced during py3 test fixing festival and is reverted here,
 - problem_site was remodeled and some buttons were moved to the new
   settings tab - tests had to be adapted.

Some tests are still failing – they are now marked xfail in strict mode,
so that anyone who fixes the underlying code will have to remove the
mark.

Change-Id: Ib9c3f681a405c8707857ef0b0f574685952f5343