(no-ticket) Gotta go fast
A major slowdown is incurred on Szkopuł when building the list of
contests that the user is permitted to see (for whatever purposes,
it is calculated for like every view or so anyway).

This patch attempts to mitigate this by NOT DOWNLOADING A WHOLE
LIST OF CONTESTS AND THEN USING IT IN SQL THROUGH `WHERE contest.id
IN ('list', 'of', 'contests', 'goes', 'here')`.

The CPU time spent should be at least an order of magnitude better,
the queries count may even be two orders of magnitude better,
as seen in in vitro tests (1410 contests, 20000 users,
25000 participants assigned randomly).

Mixins don't operate on classes verbatim, but only on constructed
objects, so the code instead assumes (and the assumption seems to work
pretty well with upstream code) that contest argument to controller
constructors is hardly used.

Change-Id: Ia235e297b046228e4770f166939e7a0fc3415cff