(SIO-2195) Implement problem statistics
This change introduces statistics calculated per-problem, as well as
per-user/problem. The statistics are cached using two helper models, and
every time a solution is submitted they are updated in constant time -
without any additional database polls. To ensure performance in
critical environments this mechanism can be disabled at will in the
settings. Any stale statistics can be recalculated with the relevant
management command. As of this change the statistics are not yet
displayed anywhere and can only be seen directly in the database.

Change-Id: I2b667cc18578c3a264e0e894062ace8b75339bbf