BUGFIX: upgrade django-celery
django-celery package is used to integrate Celery job queues into Django
admin panel and to run job worker daemons using manage.py.

The version of django-celery as pinned in setup.py before this commit is
not compatible with Django 1.10 due to the removal of optparse support
for Django management commands.

Change-Id: I00b34621dc2526535b4bd461b43b39e06543e29e