(SIO-2157) tests: Replace assertIn with assertContains
assertIn and assertNotIn are used extensively for defining test cases.
Both these functions operate on sequences (strings or bytestrings in
this case).

Django provides its own extension, assertContains, which operates on a
response object and a text type string (unicode on py2, str on py3).

Using assertContains and assertNotContains provides a more semantic way
of expressing test cases, and has a useful side effect of increasing py3
compatibility.

It can also replace another type of test cases of the form:

        assertEqual(response.content.count(x) == y)

with:

        assertContains(response, x, count=y)

where `count=` is not even needed, but I think it increases readability.

Change-Id: Idef892be1ca8eb270cbeef10e372f0f89558f1a3