(SIO-2195) Fix statistics
This changes fixes various bugs that unfortunately only became apparent
after trying the statistics on production problem/submission data.

The major problems that this addresses are:

 1. Null score_report.max_score on some older submissions as well as
    submissions imported from main. This is worked around by looking
    at tests for the problem instance and deriving max_score from there.

 2. Null submission.score and score_report.score on submissions to
    problems with no final tests, or on broken submissions. These
    are counted as submitted only.

 3. Submissions which have score > max_score, or max_score != 100.
    The score is now capped and scaled to be a percentage of max_score.

 4. Submissions with score_report.max_score correct or possible to
    calculate, but equal to 0. These are also counted as submitted only.

Added more tests to cover all 4 of these points.

Another problem was that the recalculate_statistis management command
turned out to be taking noticeably long to execute - about 3 hours on
250k valid submissions to 15k problems. After some experimenting, I
was unable to reduce this time substantially, so instead this change
ensures the management command running in the background does not
permanently block the database from normal usage, i.e. it should be
safe to leave it running in the background after enabling statistics.

Change-Id: I669e888ef3530c7cc9cce11eb013ce0594424db5