(no-ticket) Make forum test more meaningful
Before this change, test_no_forum_menu verified that forum is
not available in a contest without forum only by checking if
"Forum" appears anywhere in a dashboard of a contest that doesn't
have a forum in the controller.

The controller of choice was OIOnsiteContestController, and while
it doesn't have forum, it also makes the contests hidden to
unrelated users (such as the user used in the test).

This patch adds the user to the contest, so that it can see
the dashboard and then verify that there's no forum in it, which
is probably a more useful test.

Change-Id: Iaf0258fa4f84bce24ac030ab4f174c3ad28ad77b