(SIO-2240) Clean up problemset's main layout
This commit consists of a few small changes to the overall design:
1. Removed titles in problemset's subpages in favor of highlighted headers
2. Set navbar links to be the same ones that are on the main page
3. Removed upper pagination bar and spaced things out a little

The second one on the list also features a rework to the way the app handles navbar links.
Now it's a separate block in the main template with a new subtemplate that, using any MenuRegistry,
creates the navigation bar links.

Also removed the hardcoded Task Archive url.

Change-Id: I9030af78bd010708212545588bd6c9ba2136b9cf