SIO-2246 Add API endpoints for submiting solutions
This commit adds API endpoints for submitting solution of problems
for both contest and problemset.

It is also first example of integration of existing SIO 2 functions
with the Django Rest Framework interface. It has proper autamaticly
generated documentation and can be used as reference for adding
descriptions to path parameters or optional post fields.

I am not sure if contest app is the best choice for those endpoints.

Change-Id: Id2ce36d7a2fdca5b7598f4734b348e1fb2c36527