(SIO-2253) Create new models for storing compiler info
Added two new many to one relations representing compilers associated
with different models:

a) ContestCompiler, which represents contest-wide compiler settings. This will allow
contests to locally override compilers set for problems.

b) ProblemCompiler, which represents compilers used for this problem when no others
are specified. Instances of this are created when a new problem is uploaded
and set to values taken from DEFAULT_COMPILERS setting.

This change also increments the config version in settings to 34.

Change-Id: I7a3e0e6444211344fb08883169d1d509405e596b