(SIO-2243) Remodel OriginTags
Part 1/5 of "Remodel OriginTags for new Task Archive"

Improved the system of OriginTags to work better with the idea of a
dynamically generated task archive. The changed system works as follows:

There are three models describing a task's origin:
 - OriginTag - for general origin information, e.g. 'oi'
 - OriginInfoCategory - category specifying the OriginTag, e.g. 'year'
 - OriginInfoValue - allowed value for the category e.g. '2011'

A complete 'tag' as it would appear for example in the problemset, is
a combination of tag and value, e.g. 'oi_2011'. For this reason the
values are validated to only contain lowercase letters, numbers, and
hyphens (no underscores).
This also means that OriginInfoValues should be descriptive on their
own without the category (i.e. for Round 1, 'r1' is good, '1' is bad)

The models store additional localized values for displaying neatly to
users in their respective Localization models - OriginTagLocalization,
OriginInfoCategoryLocalization, and OriginInfoValueLocalization.
(This is needed because the values are stored in the database, not in
 the code, so it is impossible to localize them in the usual way)

Change-Id: I06459423e9441f0fd783e59702fac6f191531169