(SIO-2243) Display new OriginTags in the Problemset
Part 3/5 of "Remodel OriginTags for new Task Archive"

Added displaying the new OriginTags and OriginInfoValues in the
Problemset. The OriginInfoValues are grouped by their parent tag, and
are displayed next to it with just their value shown.

For instance in the Problemset it may look something like this:

    [pa] [2011] [r1] [a]

for OriginTag 'pa' and OriginInfoValues 'pa_2011', 'pa_r1', 'pa_a'.

Change-Id: Ifd46881d648b8f9cab4e7e0d370eb916f0a68b64