(SIO-2243) Rework the Problemset search
Part 4/5 of "Remodel OriginTags for new Task Archive"

Rebuilt the problemset search to make use of the new tags and generally
be more sensible, like using not using logical OR for every word - see
the `search_problems_in_problemset` function for details.

Removed the search hints, which will be readded in the next commit.

Replaced all of the old searching tests with ones that are a little
more compressed and less copy-paste'y.

Change-Id: Ida064249855de7f4bd65d69c4b4fe7768b9201e5