(SIO-2259) Create 'basicadmin' contest permission
This change adds a new permission 'contests.contest_basicadmin', which
is meant to be a "softer" version of the 'contests.contest_admin' -
with access to only parts of contest configuration. This is to prepare
the groundwork for limiting access in the User Contests.

Some checks for 'is_contest_admin' were left unchanged wherever the
'basicadmin' users should not have access, whereas others were changed
accordingly to 'is_contest_basicadmin' to allow these users in.

Also added the 'is_contest_admin" check on some sidebar menus, to
prevent 'basicadmin' users from seeing links to pages they don't have
access to.

Change-Id: I1a2d40a6d4e350f31b3ee3601a86f8182a05fd1d