(SIO-2259) Introduce User Contests
This change introduces User Contests, which are contests available for
any user.

An option to disable the user contests will be added by the next change.

Also fixed a bug that resulted in not displaying icons when user had
more than one menu available, as well as a weird behaviour in the
ContestAdminMixin in teachers, which was "capturing" the execution paths
belonging to other ContestAdmin mixins.

Change-Id: I9bdceea9d2d0265ef4bd0675f1c8b50e1ed4aae4