(SIO-2272) Handle user-set compilers
Use ContestCompiler and ProblemCompiler to determine what to use to
compile a particular submission. ContestCompiler overrides
ProblemCompiler if both are set. We assume ProblemCompiler is set for
every language.

If USE_LOCAL_COMPILERS is set to True, this logic is not applied,
because system-[lang] is used as a compiler anyway.

Change-Id: I77a1b57f500b8ddc12a7cd23e4a8830f1e95e8a6