SIO-2278 Navbar translations doesn't work in problemset
Translation was imported as ugettext, which was a not lazy version of translation.
Therefor, the translation was made on init, so the text once translated never changed.
Edited import to ugettext_lazy, so that the translation is lazy.

Changed jsonify functions, because they don't accept lazy translation objects.
Before jsonifying results, translated texts to current language.

Added a regression test.

What's interesting, that a bug in problems' view caused an error in szkopul view...
Probably szkopul loaded navbar view in problemset from normal sio2 view.

Change-Id: I0661b460c2cdf8fd48c8932dc25e0010e78a1e64