(no-ticket) Fix the problemset fix printing None
In my recent fix to the problemset performance issues I moved problem
tag prefetching into a template tag. Unfortunately I didn't realize that
Django would be printing the return value of the tag in the template,
even if that return value was None (which I forgot was the default
return value in Python).

Change-Id: I62913179dd2845e6529df0438e1b6d64a341b356