(no-ticket) Don't fetch visible message IDs hastily
visible_ids are used only to create a link in the top bar to the new
message if the outstanding message count is equal to 1. Therefore it is
not necessary to fetch them otherwise.

This patch speeds up all requests in contests by skipping some
unnecessarily instantly materialized database queries.

Change-Id: I73041eae78d6a56177b06d130ffffd7f3685ca1e