SIO-2286 Quiz settings have few unnessesary, but complicated options
Removed 'ignore_case' field in question model.
Made case insensitivity always set to True.
On the other hand, added it to contest controller,
so it can be overwritten for certain contests.

Also swapped points field with "is multiple choice" checkbox,
so that both checkboxes are next to each other,
and both of them are below point field.
It looks more estetic to me after made changes.

Change-Id: I47d0f1f1e663cb66efbd46e52ddf0a5ef4f56c5f