(SIO-2289) Added last submission info in Task Archive
If a user is logged in, for each problem they have attempted solving they will
see the score of their latest submission. The score is clickable and leads
to the last submission's page with full summary.
This functionality is virtually the same as in the old problemset, except now
the scores are highlighted with color.

Additionally, a user will see the proportion of their completion status
(as a percentage) for every contest they have sent at least one submission for.
(The submission needs to be successful, though the result doesn't have
to be positive; submissions that won't compile are not considered).

Change-Id: Iad59b5dd172ce149271d412a66926b589c75d03b