SIO-2288 Setting for the teacher to disable ranking in the contest
Added new configuration named RankingVisibilityConfig
in contest's settings panel. It allows the teacher
or administrator to hide ranking for regular users.
When the configuration is not present or it is set
to 'AUTO' it uses method default_can_see_ranking,
which can be overwritten in special contests.
Changes in migrations are required for previous
tests to pass, because I've added a new model.

Change-Id: I2b75b1a4ff04ca9f156be29e4bba5962d658ae69