(no-ticket) Untemplate quiz picture rendering
According to the data we collected from APM, building template context
is a surprisingly expensive operation, especially when invoked tens of
times.

During OIJ, the contestant has 2 versions of a quiz available, one for
C++, one for Python. Each of these is available in two flavours, one
that gives points and another that's for training only. A quiz consists
of 30 questions with answers and 30 without answers. Each question and
each answer has a separate template. Each one of these is rendered
completely separately, which means a template context is built under the
hood. Therefore when GETting or POSTing the submission form for OIJ
(even if it's just for a programming problem, not for a quiz), the user
has to wait like 30 seconds. That's a lot of fun.

This change, inspired by a method I saw on ProblemPackageAdmin, brings
back ye auld good string formatting in order to kick out the evil feudal
picture template lords who unfairly tax every request they see, or if
you, the Reader, are more socialist than anti-feudal, seize the means of
production and bring cheap rendering to the masses.

Change-Id: Iaa67658e55b20e48c49d660400d8b29049cff36c