(no-ticket) Exchanged the name and type of max score field in TestReport
The TestReport.test_max_score field of the IntegerField type has been
exchanged for a TestReport.test_max_score field of the IntegerScore type.
The new field's value is determined by the old field's value accordingly
to the type of the contest that the report ultimately refers to.

This change's intent is to fix the apparently illogical design where
the test reports would store the default test-specific max_score values
instead of their own, report-specific outcomes.

Change-Id: Id4b5d81957ef35f8dba48e24737f9f0db03532e1