(SIO-2049) Test report color differentiation between "OK" and "Partially OK"
Just like with the submission color, the test reports marked as "OK"
will now be highlighted with different shades of green, depending on
the fraction of the maximum score available for the test obtained
by the submission.

Test reports without defined maximum scores will always be highlighted
according only to their status, i.e. those of them that are marked
as "OK" will always receive the default, full shade of green.

Change-Id: I195ae2a7940083b53a6d77999b03ba051436703b