(SIO-2359) Problem statement link added for users in Submission View
Previously, in Submission View admins could click on the problem name
to see all solutions submitted for that problem.

Now, the problem name in this view is clickable also for regular users
and allows them to conveniently jump to the problem site or see the
problem statement, depending on the origin of the problem.

Change-Id: Ia27acc741883114ab30ae2a78be885d9283d4ad1