(no-ticket) Make questions list narrower.
Removed redundand fields "publication date" and "date".
Instead, replaces them with one field - "Appearance date" (Data ukazania).
It shows the publication date, if such exists, and if not
(it's the case when a question is asked by user), it shows creation date.

It's worth considering making the publication date no earlier than creation date.
It might be misleading, if question was created on 9:30,
but publication date is set to 9:00 and now it will be the only visible date.

Change-Id: I6073b10ea69763adc82925e27a5c1c21182ca5c1